+++ pkgs/dart_mcp_server/CHANGELOG.md
@@ -4,6 +4,7 @@
   failure.
 * Add failure reason field to analytics events so we can know why tool calls are
   failing.
+* Allow for multiple package arguments to `pub add` and `pub remove`.
 
 # 0.1.0 (Dart SDK 3.9.0)
 
@@ -54,3 +55,4 @@
 * Screenshot tool disabled until
   https://github.com/flutter/flutter/issues/170357 is resolved.
 * Add `arg_parser.dart` public library with minimal deps to be used by the dart tool.
+